refactor: Move AirVisual data classes into separate file
This commit moves the AirVisual data model classes, including AirVisualResponse,
AirVisualData, AirVisualCurrent, AirVisualWeather, and AirVisualPollution, into a
separate file named AirVisualData.kt. This makes it easier to organize the data
layer components of the app and improves the readability of the code.
