+++ app/src/main/java/com/krisna/airwise/data/models/AirVisualCurrent.kt
@@ -1,0 +1,6 @@
+package com.krisna.airwise.data.models
+
+data class AirVisualCurrent(
+    val weather: AirVisualWeather,
+    val pollution: AirVisualPollution
+)
