+++ app/src/main/java/com/krisna/airwise/data/models/AirVisualData.kt
@@ -1,0 +1,8 @@
+package com.krisna.airwise.data.models
+
+data class AirVisualData(
+    val city: String,
+    val state: String,
+    val country: String,
+    val current: AirVisualCurrent
+)
