+++ app/src/main/java/com/krisna/airwise/data/models/AirVisualPollution.kt
@@ -1,0 +1,9 @@
+package com.krisna.airwise.data.models
+
+data class AirVisualPollution(
+    val ts: String,
+    val aqius: Int,
+    val mainus: String,
+    val aqicn: Int,
+    val maincn: String
+)
