+++ app/src/main/java/com/krisna/airwise/data/models/AirVisualResponse.kt
@@ -1,0 +1,6 @@
+package com.krisna.airwise.data.models
+
+data class AirVisualResponse(
+    val status: String,
+    val data: AirVisualData
+)
